+++ examples/service/flex/plugin_configuration/page/plugin_configuration_page_example.go
@@ -1,0 +1,37 @@
+package main
+
+import (
+	"log"
+	"os"
+
+	"github.com/RJPearson94/twilio-sdk-go"
+	v1 "github.com/RJPearson94/twilio-sdk-go/service/flex/v1"
+	"github.com/RJPearson94/twilio-sdk-go/service/flex/v1/plugin_configurations"
+	"github.com/RJPearson94/twilio-sdk-go/session/credentials"
+)
+
+var flexSession *v1.Flex
+
+func init() {
+	creds, err := credentials.New(credentials.Account{
+		Sid:       os.Getenv("TWILIO_ACCOUNT_SID"),
+		AuthToken: os.Getenv("TWILIO_AUTH_TOKEN"),
+	})
+	if err != nil {
+		log.Panicf("%s", err.Error())
+	}
+
+	flexSession = twilio.NewWithCredentials(creds).Flex.V1
+}
+
+func main() {
+	resp, err := flexSession.
+		PluginConfigurations.
+		Page(&plugin_configurations.ConfigurationsPageOptions{})
+
+	if err != nil {
+		log.Panicf("%s", err.Error())
+	}
+
+	log.Printf("%v plugin configuration(s) found on page", len(resp.Configurations))
+}